Split MoveJumps class (black and white)
