+++ CheckersApplication/CheckersApplication/CheckersApplication.csproj
@@ -130,6 +130,8 @@
     <Compile Include="Detection.cs" />
     <Compile Include="CheckersPiece.cs" />
     <Compile Include="ChessBoardState.cs" />
+    <Compile Include="Moves\ForBlackMen.cs" />
+    <Compile Include="Moves\ForWhiteMen.cs" />
     <Compile Include="Moves\MovesJumps.cs" />
     <Compile Include="ToBitmapConverter.cs" />
     <Compile Include="ViewModel\MainViewModel.cs" />
